Decor: token work

+++ src/main/java/gg/projecteden/nexus/features/resourcepack/decoration/TypeConfig.java
@@ -15,7 +15,7 @@ public @interface TypeConfig {
 
 	double price() default -1;
 
-	double tokens() default -1;
+	int tokens() default -1;
 
 	Theme theme() default Theme.GENERAL;
